% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raspberry-data.R
\docType{data}
\name{ad_rd100}
\alias{ad_rd100}
\title{Arabidopsis RD100 data.}
\format{Gene data frame with  16 variables and each line represent a exonpath:
\describe{
  \item{sample}{sample name, as a string.}
  \item{sample_frag_count}{total fragment count for a sample.}
  \item{gene_id}{gene ID.}
  \item{gene_frag_count}{total fragment count for a gene.}
  \item{transcripts}{transcript ids, comma separated.}
  \item{FPKMs}{trainscript FPKM(Fragments Per Kilobase of transcript per Million mapped reads), comma separated.}
  \item{conditional_probabilities}{Conditional on a transcript, the probability of a read generated by a exonpath, comma separated.}
  \item{class_probabilities}{Probability of a read generated by a transcript, comma separated.}
  \item{path_symbol}{a unique id for a exonpath, indicating the genomic locations.}
  \item{path_count}{exonpath fragment count}
  \item{path_gc_content}{the GC content of a exonpath}
  \item{path_hexmer_entropy}{the averaged hexmer entropy fo a exonpath}
  \item{gc_stretch_0.8_20}{A indicator indicating whether the exonpath contains a stretch of DNA at least 20bp and have gc ratio > 0.8}
  \item{gc_stretch_0.8_40}{A indicator indicating whether the exonpath contains a stretch of DNA at least 40bp and have gc ratio > 0.8}
  \item{gc_stretch_0.9_20}{A indicator indicating whether the exonpath contains a stretch of DNA at least 20bp and have gc ratio > 0.9}
  \item{gc_stretch_0.9_40}{A indicator indicating whether the exonpath contains a stretch of DNA at least 40bp and have gc ratio > 0.9}
}}
\source{
\url{http://https://github.com/ruolin/raspberry}
}
\usage{
ad_rd100
}
\description{
Exonpath counts and various information for 4961 multi-isoforms genes from 6 arabidopsis thaliana samples. These 6 samples are from two groups. A dataset contains a list of 6 lists, corresponding to one sample. Each sample contains a list of genes and each gene is a data frame. These lists and tables are generated by Strawberry \url{https://github.com/ruolin/strawberry}, which is a cpp program that efficiently analyzes each RNA-Seq alignment file (BAM) and produces these tables.
}
\examples{
data("ad_rd100")
}
\keyword{datasets}
