% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{loadData}
\alias{loadData}
\title{Load data from strawberry input.}
\usage{
loadData(dir)
}
\arguments{
\item{dir}{A directory where the strawberry exonpath context output are stored.}
}
\value{
A list of of samples in Raspberry data format.
}
\description{
Load data from strawberry input.
}
\examples{
loadData("inst/extdata/AT_RD100_exonpath_context/")
}
